Вынес mapping-расширения в отдельный файл
